requirements.txt
# pip3 install -r requirements.txt
deepdiff
PyYAML
python-vagrant
ansible

# PyYAML==5.3.1
# ruamel.yaml==0.17.4
# ruamel.yaml.clib==0.2.2
# ansible==3.3.0
# ansible-base==2.10.8